Removed some code vulnerabilities
