+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## [Unreleased]  <!-- Do not delete -->
 
 ## [2024.1.0] - 2024-08-15
+
 ### Added
 - Gradle support (beta)
 - Added info link to GitHub page
