+++ src/main/kotlin/com/cquilez/pitesthelper/services/MavenService.kt
@@ -4,8 +4,6 @@ import com.cquilez.pitesthelper.model.MutationCoverageCommandData
 import com.intellij.openapi.module.Module
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.rootManager
-import com.intellij.openapi.roots.ModuleRootManager
-import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowManager
 import org.jetbrains.idea.maven.execution.MavenRunner
@@ -39,16 +37,5 @@ object MavenService {
                 }
             }
         }
-    }
-
-    fun locateSrcMainJava(module: Module): VirtualFile? {
-        val moduleRootManager = ModuleRootManager.getInstance(module)
-        val sourceRoots = moduleRootManager.sourceRoots
-        for (sourceRoot in sourceRoots) {
-            if (sourceRoot.path.endsWith("src/main/java", true)) {
-                return sourceRoot
-            }
-        }
-        return null
     }
 }
