+++ src/main/kotlin/com/cquilez/pitesthelper/services/PITestService.kt
@@ -27,7 +27,7 @@ object PITestService {
     /**
      * Collects target packages and classes in a List
      */
-    fun collectTargetCode(
+    private fun collectTargetCode(
         codeItemList: List<CodeItem>
     ): List<String> {
         val targetClassesList = mutableListOf<String>()
