+++ src/main/kotlin/com/cquilez/pitesthelper/services/UIService.kt
@@ -10,15 +10,6 @@ import com.intellij.openapi.components.Service
 class UIService {
 
     /**
-     * Show dialog only if unit test mode is disabled
-     */
-    fun showDialog(mainAction: Runnable){
-        if (!ApplicationManager.getApplication().isUnitTestMode) {
-            mainAction.run()
-        }
-    }
-
-    /**
      * Show dialog only if unit test mode is disabled or execute orElseAction
      */
     fun showDialog(mainAction: Runnable, orElseAction: Runnable){
