+++ src/test/kotlin/com/cquilez/pitesthelper/ui/components/Node.kt
@@ -2,5 +2,4 @@ package com.cquilez.pitesthelper.ui.components
 
 import com.intellij.remoterobot.fixtures.dataExtractor.RemoteText
 
-class Node(val remoteTexts: MutableList<RemoteText>, val parent: Node?, val children: MutableList<Node>) {
-}
+class Node(val remoteTexts: MutableList<RemoteText>, val parent: Node?, val children: MutableList<Node>)
