+++ src/test/kotlin/com/cquilez/pitesthelper/ui/components/ProjectTree.kt
@@ -1,4 +1,3 @@
 package com.cquilez.pitesthelper.ui.components
 
-class ProjectTree(val nodes: MutableList<Node>) {
-}
+class ProjectTree(val nodes: MutableList<Node>)
